requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
aiohttp
python-dotenv
requests